services/users/requirements.txt
Flask==1.0.2
psycopg2-binary
psycopg2==2.7.4
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.6.2
gunicorn==19.8.1
coverage==4.5.1
flake8===3.5.0
flask-debugtoolbar==0.10.1
flask-cors==3.0.6
flask-migrate==2.2.0
flask-bcrypt==0.7.1